moves non-page component out of pages folder
